Feature/refactor alberto

+++ app/src/main/java/com/skycombat/game/model/factory/EnemyFactory.kt
@@ -3,12 +3,12 @@ package com.skycombat.game.model.factory
 import com.skycombat.game.GameView
 import com.skycombat.game.model.Enemy
 /**
- * Represents an Enemy Factory
+ * Represents an Enemy Factory
  * @param scene : the gameview onto which the enemies will be spawned
  */
 class EnemyFactory(private val scene : GameView) {
     /**
-     * Generates the enemies
+     * Generates the enemies
      * @return Enemy to the scene
      * @see Player
      */
@@ -17,7 +17,7 @@ class EnemyFactory(private val scene : GameView) {
         val height = 100F;
         return Enemy(
             scene.player ,
-            (scene.getMaxWidth().toFloat() - width) / 2,
+            (scene.getMaxWidth() - width) / 2,
             100F,
             width,
             height,
